refactor(src): change components names & add export default
